requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
openai==1.35.0
python-dotenv==1.0.0
boto3==1.34.0
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]   